admin page logout functionality{api integrated} added

+++ admin-app/src/actions/auth.actions.js
@@ -82,20 +82,20 @@ export const isUserLoggedIn=()=>{
 export const signout = () => {
     return async dispatch => {
 
-       // dispatch({ type: authConstants.LOGOUT_REQUEST });
-       // const res = await axios.post(`/admin/signout`);
+       dispatch({ type: authConstants.LOGOUT_REQUEST });
+       const res = await axios.post(`/admin/signout`);
 
-       // if(res.status === 200){
+       if(res.status === 200){
             localStorage.clear();
-            dispatch({ type: authConstants.LOGOUT_REQUEST });
+            
 
-         //   dispatch({ type: authConstants.LOGOUT_SUCCESS });
-        // }else{
-        //     dispatch({
-        //         type: authConstants.LOGOUT_FAILURE,
-        //         payload: { error: res.data.error }
-        //     });
-        // }
+           dispatch({ type: authConstants.LOGOUT_SUCCESS });
+        }else{
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload: { error: res.data.error }
+            });
+        }
 
         
     }
